Only 1 Sidebar component determines what sidebar will be displayed: Trainer or Trainee
