+++ Sidebar/Sidebar.css
@@ -1,0 +1,55 @@
+section #side-bar {
+  /* z-index: 2; */
+  background-color: rgb(84, 166, 177);
+  /* height: 100vh;
+    width: 150px; */
+  position: fixed;
+  top: 0;
+  left: 0;
+}
+.visible {
+  z-index: 2;
+  background-color: rgb(239, 76, 76);
+  height: 100vh;
+  width: 300px;
+  position: fixed;
+  top: 0;
+  left: 0;
+  transition: left 0.5s linear;
+}
+/* .button
+{
+    background-color: rgb(124, 146, 146);
+    border-color:  rgb(124, 146, 146);
+    border-style: solid;
+} */
+#sidebar#visible {
+  width: 300px;
+  transition: left 0.5s linear;
+}
+.sidebar-btn {
+  vertical-align: right;
+  width: 20px;
+  height: 15px;
+  cursor: pointer;
+  margin: 20px;
+  float: left;
+}
+.side-bar-1st {
+  margin-top: 60px;
+  padding: 16px;
+  font-size: 30px;
+  border: 1px solid gray;
+  text-align: center;
+  text-decoration: "none";
+}
+.side-bar-1st.hover {
+  color: white;
+}
+.side-bar-data {
+  padding: 16px;
+  font-size: 30px;
+  border: 1px solid gray;
+  text-align: center;
+  text-decoration: "none";
+}
